changed type to decimal for product

+++ server/db/models/Products.js
@@ -18,11 +18,11 @@ const Product = db.define('product', {
     allowNull: false,
   },
   lat: {
-    type: Sequelize.INTEGER,
+    type: Sequelize.DECIMAL,
     allowNull: false,
   },
   lng: {
-    type: Sequelize.INTEGER,
+    type: Sequelize.DECIMAL,
     allowNull: false,
   },
   votes: {
@@ -30,7 +30,8 @@ const Product = db.define('product', {
   },
   imageUrl: {
     type: Sequelize.TEXT,
-    defaultValue: "https://cdn.pixabay.com/photo/2016/03/22/04/23/map-1272165_960_720.png",
+    defaultValue:
+      'https://cdn.pixabay.com/photo/2016/03/22/04/23/map-1272165_960_720.png',
   },
   isAvailable: {
     type: Sequelize.BOOLEAN,
